Remove ~/src/repos/github.com/martine/ from PATH.
That is not good when installing on a new totaly fresh system.

With this change we will end up using the version from ~/depot_tools,
which is not bad at all.

The other options if we didn't have depot_tools on PATH would be to have
a copy of the binary in /usr/local/bin, which I also have. So that is
good enough.

Signed-off-by: Thiago Farina <tfarina@chromium.org>