fix comment
rxvt-unicode provides the bindings Ctrl-Meta-c and <Ctrl-Meta-v> to interact with the CLIPBOARD selection.
The first binding causes the value of the internal selection to be copied to the CLIPBOARD selection,
while the second binding causes the value of the CLIPBOARD selection to be inserted.

Also the comment about not using Perl is wrong, they use perl extensions to interact with the clipboard.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>