update the image viewer I'm using
Most of the time I have been using GNOME Image Viewer (3.18.2) also known as
Eye of Gnome (eog), because that is what opens when double clicking an image
on Thunar.

Rarely I have the need to see an image or look at an image when working at the
terminal emulator.

But now, I'm switching from the gnome app to the xfce app known as ristretto.

It was installed with the following command line:

$ sudo apt install ristretto

The following command gave the version installed:

$ ristretto --version
ristretto 0.8.0

It is very lightweight and looks better than eog.

http://goodies.xfce.org/projects/applications/ristretto

Signed-off-by: Thiago Farina <tfransosi@gmail.com>