update document viewer
I'm using Evince, which is GNOME Document Viewer.

The version was found with the following command:

$ evince --version
GNOME Document Viewer 3.18.2

I'm not using zathura, because most of the time, the pdf documents are opened
when browsing from Thunar, so it is more convinient to let it open a GUI
application than to go to the terminal, cd to the directory and open it with
zathura.
GNOME Document Viewer 3.18.2

The website for this application is https://wiki.gnome.org/Apps/Evince.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>