urxvt: add settings to change the shape of the cursor
By turning them to true, the cursor changes from a block to a very small blinking line. While it may be good in the terminal prompt, it is hard to find and
see when coding in emacs editor.

The block shape is much easier to see while in emacs, hence I'm keeping this
false for now.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>