add information about init system
Systemd provides a system and service manager that runs as PID 1 and starts
the rest of the system. The kernel initiates the init process, which startsthe parent process. The parent of all Linux processes is Systemd, which replaces
the old SysVinit process.

I ran the following command line to get its version:

$ systemctl --version

The following also works:

$ sudo apt-cache show systemd | grep -i "^version"

https://www.nixcraft.com/t/how-to-find-systemd-version-on-linux/3848

It is the systemd that starts the lightdm display manager. That was found
with the following command:

$ cat /etc/systemd/system/display-manager.service

It has a line ExecStart=/usr/sbin/lightdm, which shows that it executes
lightdm.

More information about the boot process can be found at https://www.baeldung.com/linux/boot-process.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>