merge urxvt file into terminal/urxvt/profiles/basic
That way I can start working in a configuration for the xterm terminal.
Otherwise, it would be very difficult or even impossible.

That completes the reorganization of the .Xresources.d directory into a layout
that can accomodate different terminals and their configurations. It is also
more flexible to configurate other X programs.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>