create a symbolic link to urxvt's basic file
From <https://linuxhandbook.com/symbolic-link-linux/>:
What is a symbolic link?

"A symbolic link, also known as a symlink or a soft link, is a special type of file that simply points to another file or directory just like shortcuts in Windows. Creating symbolic link is like creating alias to an actual file.

If you try to access the symbolic link, you actually access the target file to which the symlink points to. Changes performed on the content of the link file change the content of the actual target file."

The link was created with the following command:

$ ln -s default basic

To verify it, I ran:

$ ls -l default

And also:

$ realpath default

The idea for using a symbolic link was taken from <https://github.com/rafi/etc-skel/blob/master/.config/bash/themes/current>.

That way I don't need to edit the .Xresources file again. I can just change
the symbolic link to point to another configuration file, rerun xrdb, and it
is done.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>