set up utmpInhibit resource option
It is false by default because we want to write record entry into the system
log file. This utmp file allows to discover information about who is using
the system. We want that and we want to go with the default setting, so we keep
that way until there is a good reason to change it.

For more information on utmp see man utmp and/or man urxvt.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>