set up pointerBlankDelay resource option
It specifies number of seconds before blanking the pointer. Since we don't
blank the pointer, i.e. we like and keep it always visible, that setting
has no effect what so ever. Anyway, we set it up to the default value, in
case we change our minds in the future.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>