set up scrollTtyOutput resource option
This is simply a setting to "Scroll on output". I like this behavior, hence it
is set to true. urxvt's manual page does not say what is the default value, so
this is fine.

Basically, this option controls whether the terminal (tty) will scroll down
automatically whenever it receives output from commands.

Note that this seems to be also the default behavior in the Xfce Terminal.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>