set up scrollTtyKeypress resource option
When you press any key on the keyboard it will scroll down the terminal
window to the command line prompt. The manual page for urxvt does not say
what is the default. So we are going with true, because that is the behavior
we want and like anyway. It is also the way Xfce Terminal is configured, so
that is good enough.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>