set up borderLess resource option
This option sets MWM hints to request a borderless window. If honoured by the
window manager, the terminal will not have borders around its window.

The way I tested this was setting this option to true and reloading the
configuration. Under Xfce window manager, when I opened a new window, it didn't
have borders around it. So xfwm honors this setting.

Since I want to be able to move the window around, resize it, maximize and
minimize it, I'm setting it to false.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>