set up override-redirect resource option
We don't need to enable or set to true this resource option, because the
window manager I use (Xfce) supports MWM hints (unlike kwin, for example).

According to urxvt's revision changes, it is necessary to enable this option
for kwin window manager.

For now, it works fine with the false value for the Xfce window manager.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>