remap copy and paste keyboard shortcuts in urxvt
I have switched to Shift+Ctrl+C for Copy and Shift+Ctrl+V for Paste,
and I have been using them for some time now that they are in my muscle
memory.

I changed them from the old Alt+C and Alt+V, because the Shift+Ctrl+
is what Xfce Terminal and Gnome Terminal use.

I found them to be a bit more ergnomic as well, so I'm sticking with it.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>