switch main font used in urxvt to Inconsolata
This is a font very similar to Consolas, which is
the reason that lead me to try it.

I have been using it for a couple of months now. It
looks good, it is better than Menlo in some points,
but has some minor issues.

Screenshot taken with xfce4-screenshooter added.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>