document the shortcut used to open the start menu
Alt+F1 is the default shortcut used to open the Start menu on Xfce.

It calls xfce4-popup-applicationsmenu.

To see or change it, go to Start menu -> Settings -> Keyboard -> Application
Shortcuts.

The next step is to use xsuperkey or xcape to map it to the Super key, to
be able to open it using the Super key without messing with the other
shortcuts that use Super.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>