add grub configuration file
This adds a copy of the origin grub configuration file,
because I'm going to make a change to display boot messages
on the screen when booting up [1].

So, before I mess up with it, let's have a backup.

This file was copied with the following command line:

$ cp /etc/default/grub ~/dotfiles/grub

As I decided not to symlink it, in order to keep this copy
updated we have to copy the original file whenever we make changes to
it to the dotfiles repo.

Refs:
https://askubuntu.com/questions/25022/how-to-enable-boot-messages-to-be-printed-on-screen-during-boot-up

Signed-off-by: Thiago Farina <tfransosi@gmail.com>