enable boot messages to be printed on the screen when booting up the system
To enable these messages and see them on the screen when the system is booting
up I had to remove quiet and splash parameters. Without passing this parameters
to linux, the messages are printed on the screen.

I edited the file with the following command:

$ sudo emacs -nw /etc/default/grub

After that, to make sure the changes take effect, I had to run the following
command:

$ sudo update-grub

Refs:
https://askubuntu.com/questions/25022/how-to-enable-boot-messages-to-be-printed-on-screen-during-boot-up

Signed-off-by: Thiago Farina <tfransosi@gmail.com>