add a copy of Clearlooks-Phenix theme directory
This adds a copy of Clearlooks-Phenix theme directory, which I installed
from apt package manager with: sudo apt install clealooks-phenix-theme.

It is a small directory, it contains 143 items and just 232.2 kB.

Unlike the version I archived from the git repository, it does not contain
COPYING, doc/, LISEZ-MOI.md, openbox-3/, README.md, src/ and wallpapers/.
That is why it is smaller.

It was copied here with the following commands:

$ cd ~/dotfiles/themes
$ cp -r /usr/share/themes/Clearlooks-Phenix .

Signed-off-by: Thiago Farina <tfransosi@gmail.com>