display GRUB menu when booting up
This makes the bootloader menu always appear after the POST screen (BIOS load).

Otherwise, I have to press and hold Shift when it is loading.

I set up a ruge timeout (30 seconds), so there is no rush and it allows plently
of time to read everything the menu says.

As always, to update this copy of grub file, I ran the following command:

$ sudo cp /etc/default/grub ~/dotfiles/grub

To apply the changes made and make the change permanent, I ran:

$ sudo update-grub

Refs:
https://askubuntu.com/questions/16042/how-to-get-to-the-grub-menu-at-boot-time

Signed-off-by: Thiago Farina <tfransosi@gmail.com>