set cutchars resource option for urxvt
I have been using this for a while now without any problems I could notice
so far.

I have used the same value as the CUTCHARS macro in [1].

This indicates which characters should be considered when double clicking.

It differs from Xfce Terminal, which defines it as the following:
-A-Za-z0-9,./?%&#:_=+@~

But for now, this one is working fine for me.

Refs:
[1] - https://github.com/exg/rxvt-unicode/blob/c149b9460b08630014a3ee44737e0b42cfca7035/src/feature.h#L349

Signed-off-by: Thiago Farina <tfransosi@gmail.com>