do not set title resource option to empty string
If we set title like this:

URxvt.title:

Then the urxvt window will have an icon but no title. urxvt will not get its
name from the basename of full path in the argv[0].

Also, in Xfce panel the window will be called "Untitled window", but we don't
want that, so let's not set it to an empty string.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>