set borderColor resource option
This paints the internal border (which I use with 2 pixels) of urxvt with
green color. This shows a thin line around the client area of the window.

I think it is funny and it seems to expose a bug (maybe in urxvt) when the
window is maximized. When that happens the bottom line and right line have
a bigger size than 2 pixels. Weird.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>