display the Start menu on Ctrl+Esc
This changes the keyboard shortcut used to display the Start
menu (xfce-popup-applicationsmenu) from Alt+F1 to Ctrl+Esc.

Windows 98 uses Ctrl+Esc to display the Start menu, so I decided
to use the same shortcut to facilitate the back and forth between
the two systems.

I swapped the keyboard shortcuts between xfce-popup-applicationsmenu
and xfdesktop --menu, as for xfce-popup-applicationsmenu the default
shortcut was Alt+F1 and for xfdesktop --menu was Ctrl+Esc.

To make the change, on Xfce, select Start menu -> Settings -> Keyboard ->
Application Shortcuts.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>