set jumpScroll resource option
This resource option specifies whether or not jump scroll should be used. If
it is false, urxvt will use smooth scrolling. I found that for my needs,
jump scroll behavior will suffice. If I notice anything weird from using it,
then I can set this to false. I'm also setting it to true as "true" is the
default value in xterm. The documentation says it corresponds to the VT102
DECSCLM private mode.

On aixterm, the default is false according to the documentation. dtterm
also have this resource option, but the documentation does not says what the
default value is.

urxvt does the following on init_command function:
if (!option (Opt_jumpScroll))
    priv_modes |= PrivMode_smoothScroll;

Refs:
https://man.archlinux.org/man/urxvt.1
https://www.x.org/releases/X11R6.7.0/doc/xterm.1.html
https://www.ibm.com/docs/en/aix/7.2?topic=aixterm-command
https://cdesktopenv.sourceforge.io/man1/dtterm.html
https://github.com/exg/rxvt-unicode/blob/c149b9460b08630014a3ee44737e0b42cfca7035/src/init.C#L1112

Signed-off-by: Thiago Farina <tfransosi@gmail.com>