start configuring xterm
I'm using the file from meribold's dotfiles [1] as a start point. My plan
is to switch from urxvt to xterm, once I'm confortable using it.

To create the symbolic link, run:

>cd x11/.Xresources.d/terminal/xterm/profiles/
>ln -s basic default

To remove the symbolic link, run:

> rm default

Or

>unlink default

To verify it was created or removed, run:

>ls -l

To see the real path to where "default" points to, run:

>realpath default

Refs:
[1] - https://github.com/meribold/dotfiles/blob/cb0e502b0a2746814431c3b8d97a8c4379c90442/home/xresources/xterm

Signed-off-by: Thiago Farina <tfransosi@gmail.com>