document the keyboard shortcut used to open the task manager
This adds a keyboard shortcut to open the task manager (lxtask).

To add a new keyboard shortcut in Xfce select Start menu -> Settings ->
Keyboard -> Application Shortcuts and click on Add button.

LXDE Task Manager (lxtask) is the equivalent on Linux to the Windows
Task Manager.

This keyboard shortcut is documented here [1], works on Windows XP,
but I couldn't find it documented on Windows XP Help documentation.

Refs:
[1] - https://support.microsoft.com/en-us/windows/keyboard-shortcuts-in-windows-dcc61a57-8ff0-cffe-9796-cb9706c75eec

Signed-off-by: Thiago Farina <tfransosi@gmail.com>