switch urxvt font to Consolas
I have been using it for a while now. It is pretty good and looks
very well in this configuration on urxvt. Even though it looks good,
it has some characters I don't like, as: g,r,l,m,?,@ and $.

The g character could be simpler. The arm of the r could be staright. The l
looks like 1. The legs of the m character are too tight. The shape of the
? could be rounded. The strike line in in the f character could be higher.

Besides that, I like the overall shape and design of the other characters.

The added screenshot was taken with xfce4-screenshooter.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>