add link to the font article that helped me configure urxvt
This archlinux.org article helped me configure the fonts on urxvt.

This particular set of settings led to the best look and feel on my
lcd screen.

It is so, that I have Xfce set exactly the same. I even tried other
combinations, but all of them led to blurry and bad looking fonts.

To configure the rendering of fonts in Xfce, go to Start menu -> Appearance
-> Fonts.

On the terminal, I can retrieve these values with the following commands:

> xfconf-query -c xsettings -p /Xft/Antialias
1
> xfconf-query -c xsettings -p /Xft/Hinting
-1
> xfconf-query -c xsettings -p /Xft/HintStyle
hintslight
> xfconf-query -c xsettings -p /Xft/RGBA
rgb
> xfconf-query -c xsettings -p /Xft/DPI
96
> xfconf-query -c xsettings -p /Xft/Lcdfilter
lcddefault

Refs:
https://wiki.archlinux.org/title/Font_configuration#Applications_without_Fontconfig_support

Signed-off-by: Thiago Farina <tfransosi@gmail.com>