document another keyboard shortcut I use to open the Start menu
Until now, I couldn't use Super as modifier or keyboard shortcut to open
the Start menu. It simply does not work and if we try to add and use it,
it breaks the other combinations like Super + R or Super + E.

To workaround this Xfce "bug", I had to use ksuperkey. It allows us
to open the Start menu using the Super key (also known as "Meta" or
"the Windows key"). If we hold down the Super key it will still
act as a modifier key, allowing us to use it for other keyboard
shortcuts.

To build ksuperkey, just run the following commands:

> git clone https://github.com/hanschen/ksuperkey.git
> cd ksuperkey
> make

For installing it system-wide, run:

> sudo make install

Now the tricky part. We have to make Xfce run ksuperkey everytime the
system starts. To do that, we go to Start menu -> Settings ->
Session and Startup -> Application Autostart. There we click on
"Add" button and give it a name, a short description and the command
we want to run.

Since I already have Ctrl+Esc configured to open the Start menu, we are
going to use that. So we enter the following on the Command input:

> ksuperkey -e "Super_L=Control_|Escape"

Now when we press and release the Super key, the Start menu opens up.

The attached autostart file contains all this information.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>