configure cursorColor together with the other color options
This is how Windows Terminal does when configuring the color schemes.

Since it is a color and it is integral to the look and feel of
the terminal, it makes sense to keep them together.

I plan to change this color to black to make it match the color of
the prompt symbol.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>