x11: change cursor color to black
This grey color dates back to commit f9c72a5cd0f5 ("x11: add color scheme inspired on Basic Profile from MacOS Terminal").
So this grey color is from Apple Terminal Basic color scheme.

The cursor color for Vintage and Campbell color schemes from
Windows Terminal is #FFFFFF [0,1], which is white in a black background.

I'm changing it to black, so it matches with these color schemes and
also with the color of my prompt symbol.

Refs:
[0] - https://github.com/microsoft/terminal/blob/4bcdd7a8444ebb46f968932f9975e76320250f85/src/cascadia/TerminalSettingsModel/defaults.json#L177
[1] - https://github.com/microsoft/terminal/blob/4bcdd7a8444ebb46f968932f9975e76320250f85/src/cascadia/TerminalSettingsModel/defaults.json#L133

Signed-off-by: Thiago Farina <tfransosi@gmail.com>