x11: set the color values of foreground and background directly
This dates back to commit 2a9eb4714353 ("give names to background and foreground colors").

At that time, I thought it was a good idea, and use it. But I ended
up not using the idea and now looks like it makes more sense to
just put the values directly there instead of having this indirection.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>