urxvt: change font to Ubuntu Mono
I have been using this font instead of Consolas for
more than a month now and I'm really enjoying it.

The characters I don't like on Consolas, I like how
they are on Ubuntu Mono. They look great. But I don't
like the dollar sign ($) and the at sign (@).

Another thing I like on Ubuntu Mono is how the strike on
f letter aligns perfect with the strike on t letters. We
can see that on the screenshot.

Also, the line height is better in this font, they don't
touch the green line (border). The screenshot also shows
that.

Overall, I like more all the characters in this font than
the ones in Consolas.

I changed the font size to 15 pixels, because after some
testing, it looked better in this size.

I made a copy of the Ubuntu Mono font in my fonts repository.
The .ttf files were added at commit 827ec200c40d ("add Ubuntu Mono font").

https://github.com/tfarina/fonts/commit/827ec200c40d5acf351c3e9ad04cbfe547d8e7a4

Signed-off-by: Thiago Farina <tfransosi@gmail.com>