urxvt: change font size to 12 points (pt)
I was setting the font size in pixels and the previous size was 15px.
With that size, Ubuntu Mono looked great on urxvt. But I was not completly
satisfied, I was feeling it was a little bit small, very subtle I know.

So, I picked the size 12 from Windows Terminal [0,1,2] and tried it to see how
it would look like. And it worked great. The font got bigger, the
characters were bigger, but the width didn't increate, which was a great
thing because the overall width of the terminal didn't increase too.

I changed from defining it pixels to point, because according to the
documentation, the point size is automatically converted to the correct
pixel size based on the resolution of my monitor and I use dpi with a
value of 96 (xdpyinfo | grep dots).

The formula according to W3C [4,5] is:

1px = 1/96th of 1in
1pt = 1/72th of 1in

So, from that we have:
96px = 72pt
1px = 72pt / 96

To convert pt to px, we just have to multiply pt value to (72pt / 96):
px = pt * ( 72pt / 96 )

Refs:
[0] - https://learn.microsoft.com/en-us/windows/terminal/customize-settings/profile-appearance
[1] - https://github.com/microsoft/terminal/blob/4bcdd7a8444ebb46f968932f9975e76320250f85/src/cascadia/TerminalSettingsModel/defaults.json#L67
[2] - https://www.tenforums.com/tutorials/179097-how-change-font-size-windows-terminal-profile-windows-10-a.html
[3] - https://www.w3.org/TR/css3-values/#absolute-lengths
[4] - https://pixelsconverter.com/pt-to-px

Signed-off-by: Thiago Farina <tfransosi@gmail.com>