urxvt: change cursor shape to underscore
I was using a filled box, but the shape of the cursor in
Windows XP Command Prompt is like a block or a box. It is
way bigger than a underscore (_), but urxvt does not have
anything similar, the closest to it is the underscore.

I was reluctant to switch to a underscore cursor, because of Emacs.
Sometimes it is hard to find it there, but I will get used to it.

Windows Terminal has support for many cursor shapes and they have
support for the Command Prompt cursor, they call it "vintage". The
one they use by default is called "bar" (|) though.

Refs:
https://learn.microsoft.com/en-us/windows/terminal/customize-settings/profile-appearance#cursor

Signed-off-by: Thiago Farina <tfransosi@gmail.com>