Revert "use rgb:<red>/<green>/<blue> syntax for defining color values"
This reverts commit 6ea1d9b6794986ca22b4e4e030528036efca29ba.

I'm reverting the changes made in this commit for some reasons.

The main reason is that the hex format #RRGGBB is much more common.
It is pretty much used everywhere.

Second reason is because when you search for rgb(192,192,192) or any
other rgb combination on DuckDuckGo or Google search engines, they
give you back the color in hex format and vice-versa. Not in
this rgb:RR/GG/BB format.

So to make it easier to go back and forth, to copy and paste values
from other places, I'm changing this back to the hex #RRGGBB format.

gcolor2 tool also gives you the color in hex # if you input the red,
green and blue values.

The color scheme values in the Windows Terminal are also defined in
hex # format.
https://github.com/microsoft/terminal/blob/4bcdd7a8444ebb46f968932f9975e76320250f85/src/cascadia/TerminalSettingsModel/defaults.json#L178

So...

Signed-off-by: Thiago Farina <tfransosi@gmail.com>