urxvt: rewrite rgb values in a simpler way
When I search for a hex color with DuckDuckGo search engine,
for example, #FFFFFF, it returns the rgb value as RGBA(255, 255, 255, 1).

On Google, it returns like the rgb value like this:

RGB
255, 255, 255

Yandex does not have this color picker feature.

Bing returned:

sRGB (r, g, b)  (255, 255, 255)

So, adding the space between the values and removing the rgb and parenteses
makes my job of copying from Firefox and pasting on Emacs much easier, and
saves a bunch of time as I don't have to edit the text pasted to remove
the spaces between the commas if I were to keep the old style.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>