urxvt: remove grey50 color
We used this color for the cursor and it was from Apple Terminal
Basic color scheme.

Since commit d5dcdad65a4d ("x11: change cursor color to black"),
it became unused. If we want to change the cursor color again,
we can just put the new value in hex # format directly there.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>