urxvt: put resource option descriptions together with their respective options
This is a cleaner way to organize this file. When changing an option, if
the description is right there above and not somewhere very far away, you
can read it right away to know what it does and which values it can take.
Even if you come to remove an option, if the description is right there,
you won't forget to remove it. Which is something that could happen if the
description is not attached to its resource option.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>