urxvt: write termName option description in third person singular
The best way to write the description for options is in third person
singular, using "it" pronoun, and in simple present tense.

Why? Because these resource options are not commands or actions, and
thus are not verbs. When writing the description in third person singular
in simple present it sounds like "this option does this", so it is
like reading as "termName specifies ..." or "it specifies" or "it controls".
This way the description is telling what the options does, what is its
behavior or effect, as opposed to be giving a command.

A good example of writing in this style is the XTerm manual. Many option
descriptions are writing in simple present tense (third person singular).

Refs:
https://www.grammar-monster.com/glossary/third_person.htm
https://www.grammar-monster.com/glossary/simple_present_tense.htm

Signed-off-by: Thiago Farina <tfransosi@gmail.com>