urxvt: simplify header sections
The new way to write a section is:

! Section title
! =============

Or:

! Sub-section title
! ------------------

Use the equal sign (=) for main levels and hyphens (-) for
secondary levels.

This syntax was borrowed from Markdown [0].

The main advantage over the old way of writing !---- above
and !---- below is that the old way you never get it right
in terms of length or width, because if you change the size
of the font, then it messes up everything. With this new
style, you don't have this problem anymore. It is nice and
clean.

Also we should have only one level per section in most cases.
No more sublevels.

Refs:
[0] - https://www.w3schools.io/file/markdown-headings/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>