urxvt: document which type the option value accepts
The advantage of having this "Accepts: ..." line in the description,
right above the resource option, is that with it I don't have to open
the urxvt manual and search for this information. Nor I have to go
to the end of the file and look at the list there.

The end result is that this saves a lot of time.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>