urxvt: put document title on the left
This is for the same reasons we changed the headings
to a simple style based on Markdown.

If the default width of the terminal changes, this
center alignment will look wrong. If the size of the
font changes, same thing. So, we will never get a
perfectly centralized text there. Moving the text
to the left side will avoid all these problems.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>