add Thunderbird as an alternative to Claws Mail
I have been searching for an alternative to Claws Mail email client because
of its integration with Gmail or lack of it. As of May 30, 2022, Google
no longer allows Less secure apps access for Gmail accounts, thus OAuth
is required. And the documentation for configuring OAuth on Claws Mail
was not good or easy to understand.

I have considered GNOME Evolution [0], but it also lacks good documentation
on this problem too.

Mozilla Thunderbird on other hand, has excelent documentation on this
issue at [1].

For this reason, I'm considering migrating from Claws Mail to
Mozilla Thunderbird [2], as I use Gmail as my primary email service.

The version I have already installed on my system is 68.10.0:

> thunderbird --version
Thunderbird 68.10.0

The documentation says if I'm using Google supplied "app passwords",
I need version 91.8.0 or newer for Thunderbird to make the conversion
automatically for me.

Let's see.

Refs:
[0] - https://gitlab.gnome.org/GNOME/evolution/-/tree/master
[1] - https://support.mozilla.org/en-US/kb/thunderbird-and-gmail
[2] - https://www.thunderbird.net/en-US/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>