x11: retire themes directory
They might apply to all applications that use standard Xresources, but we
are mainly working and focusing on urxvt for now.

I decided to move these color scheme files into a new "colorschemes"
directory, under urxvt. This is the new structure I'm thinking to use,
as it is more similar to what Windows Terminal does.

To be able to use them and switch color schemes easily, we need a way
to include the color scheme file from basic.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>