urxvt: tell in the description what the default value is
Currently, we don't have this information in any of the
resource option descriptions, which makes harder and time
consuming to know what the default value is for that resource
option. Either I have to go back to the terminal and run man,
or look up online.

The good thing of having the default value right there, is that
it allows me to know if I changed it or not.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>