add screen locker info to README
I use XScreenSaver as my screen locker for now, as it is getting the job
done.

The command to lock the screen is xflock4, which is a shell script that
calls xscreensaver-command -lock. This command is run when I press Super+L,
because I associated it with this keyboard shortcut on Xfce through the
Keyboard properties dialog.

Unfortunatelly, this script has undergo some changes I really don't like,
as you can see from the commit history at [0]. So, when upgrading to a new
version, we have to look out and remember about this.

Refs:
[0] - https://gitlab.xfce.org/xfce/xfce4-session/-/blob/master/scripts/xflock4

Signed-off-by: Thiago Farina <tfransosi@gmail.com>