update window manager theme info in README file
I added a copy of Chicago95 theme at commit ce11b1978ead ("add
a copy of Chicago95 theme"), but I forgot to update the README
file.

I was probably using it for some time before I made the backup
of the theme.

On Xfce, the Window Manager theme can be changed at:

- Start menu > Settings > Window Manager > Style

This information can also be retrieved in the command line:

> xfconf-query -c xfwm4 -p /general/theme
Chicago95

For GNOME, you can retrieve it running:

> gsettings get org.gnome.desktop.wm.preferences theme
'Ambiance'

Which seems to be the same as the GTK3 theme.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>