backup .xscreensaver file
To test and do this patch, I ran a few commands, just to make
sure I knew what was going on. And to know what stow would do,
or is doing under the hood.

So, first was:

$ cd ~/dotfiles
$ mkdir xscreensaver
$ cp ~/.xscreensaver ./xscreensaver

To create a backup (or copy) of the original .xscreensaver file.

Then:

$ mv ~/.xscreensaver ~/.xscreensaver.bak

To get it away and be able to create the symbolic link (symlink),
while keeping a safe copy just in case things go wrong.

$ cd ~  # or cd .. to go back to the $HOME directory
$ ln -s dotfiles/xscreensaver/.xscreensaver .xscreensaver
To actually create the symbolic link (symlink).

To verify it worked and was right:

$ ls -l .xscreensaver

To remove the link:

$ unlink .xscreensaver

Then, back to dotfiles directory:

$ cd dotfiles
$ stow --simulate --verbose=3 --target "$HOME" -S xscreensaver
To make sure stow would do the right thing.

After checking a couple of times that stow would do the right thing:

$ stow --verbose=3 --target "$HOME" -S xscreensaver

Signed-off-by: Thiago Farina <tfransosi@gmail.com>