show the full name of the day of the week in the clock
Previously, we were not showing the name of the day in the clock.

But I think it is a nice information to display. So, let's add it
to the clock's format string.

As a bonus info for the future, the clock-format string is a
strftime string, so it accepts strftime specifiers and the %A came
from there.

Just look at the manual with the following command to see all
the specifiers you can use:

>man 3 strftime

%A  The full name of the day of the week according  to  the  current locale.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>