urxvt: document the default value for borderLess option
Currently, this option is set to false, meaning we want a
window frame with borders, like a normal window.

I could not find the default value documented anywhere in the
entries of the manual or the source code at
http://cvs.schmorp.de/rxvt-unicode/src/.

But it is very unlikely the default value of this option would
be true.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>