add a copy of Bluecurve cursor theme
Why? Because it is slightly better than the DMZ-White
cursor theme used in Ubuntu 9.04.

Its overall look is more similar to the cursors in Chicago95,
so I prefer it over DMZ-White and I'm actually using it right now.

I copied this cursor theme from Fedora 11 (Leonidas), which is
from 2009-06-09, because it came with GNOME 2.26 (same as Ubuntu 9.04).

So, let's make a backup of this cursor theme.

https://archives.fedoraproject.org/pub/archive/fedora/linux/releases/11/Live/i686/Fedora-11-i686-Live.iso

Signed-off-by: Thiago Farina <tfransosi@gmail.com>