open xfce4-appfinder in normal mode
When opened in collapsed mode, it is pretty much useless,
I have to expand it in other to tab through the results.

So, to avoid that, let's open it normally always. Bonus
point is that this is the same way my Super+R shortcut
opens it and also the launcher in the panel.

The original file lives in /usr/share/applications/ directory.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>