themes: make icon label background transparent in Clearlooks-Modified theme
Previously, label-alpha was set to 255, which makes the background of the
desktop icon label totally opaque. This is because the alpha channel is
composed of 256 levels of transparency, where 0 means total transparency,
and 255 is totally opaque.

Since we don't want any background painted at all, we now set label-alpha
to 0 to make it totally transparent.

With this change, we can now remove the base settings.

More information on alpha channel can be found at https://rgbacolorpicker.com/.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>