themes: stop setting base in Clearlooks-Phenix-Modified theme
It only made sense when label-alpha was set to 255 and the
background completely opaque. That is why we could apply
a color to the background of the icon label. Now that it is
completely transparent it does not make sense anymore.

So, let's remove this setting of teal background color, which
came from Chicago95 theme.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>