bash: add a TODO to get prompt fixed on tty
By default, Ubuntu has 7 tty's.

Up until Ubuntu 17.10: 1-6 are command line only and 7 runs
the X session (normal desktop).

On Ubuntu 17.10 and newer: It's GUI login screen on 1, GUI
desktop on 2 and command lines on 3-7.

To access them, use the following keyboard shortcut:

Ctrl + Alt + F1 (or F3 after 17.10).

Changing F1 to F1-F6 to access the terminal that you want.

To get back to GUI session (normal desktop):

Ctrl + Alt + F7 (or F2 after 17.10).

See https://askubuntu.com/questions/66195/what-is-a-tty-and-how-do-i-access-a-tty

This is a reminder to get a prompt without color on tty, because
there the background is black and my path color is black, so with
this setting I can't see it there. Look how it is done at .bashrc.
It has the answer for this.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>