backup xfce4 xfwm4 and xsettings configuration
Now that we have all the elements in place we can make a backup of these
two most important Xfce's configuration files.

The xsettings.xml file contains the theme name (used by gtk), the icon theme
name, the xft settings (dpi, antialias, hinting, etc), the font name and
cursor theme name.

These settings can be retrieved on the command line with xfconf-query.

For example, to retrieve the value of ThemeName property of the xsettings
channel, run:

$ xfconf-query -c xsettings -p /Net/ThemeName

To retrieve the value of HintStyle property of the xsettings channel, run:

$ xfconf-query -c xsettings -p /Xft/HintStyle

And so on.

The xfwm4.xml file contains the theme name (used by the window manager), the
title font used in the titlebar of the window frame, and so.

The same way, the values can be retrieved with xfconf-query.

For example, to retrieve the window manager theme, run:

$ xfconf-query -c xfwm4 -p /general/theme

To retrieve the window manager title font, run:

$ xfconf-query -c xfwm4 -p /general/title_font

And so on.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>