change the hostname of my main machine to lp2
The previous name was apollo, which is the name
of a God in the Greek mythology. I don't recall
the reason I chose this name.

The new name is lp2, which is shorter and looks
like more technical.

It is an acronym for: laptop pentium dual-core

And kind of makes more sense and creates a pattern
for naming other machines as this one is a Gateway
Laptop, equipped with an Intel Pentium Dual-Core Mobile T2060
processor.

To change the hostname permanently, I edited the file
/etc/hostname and rebooted.

To fix the sudo warning "sudo: unable to resolve host lp2",
I edited to the file /etc/hosts.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>