bash: fix command prompt in tty for linux terminal
My system has 6 tty going from tty1 up to tty6 that can
be accessed by pressing Ctrl+Alt+F1 (until Ctrl+Alt+F6).
To return to GUI desktop you have to press Ctrl+Alt+F7.

The issue is that this tty has a black background and I
set the working path to a black color, which makes the
path invisible.

To fix that we have to check the $TERM variable and make
a PS1 string without colors, because there this $TERM
variable returns "linux".

The .bashrc file has the fix for this, so I just copied
from there and adapted to my needs.

I broke this when I did commit 84da8720c188 ("bash: change the
color of path to black").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>